refactor(io): split the 3.7k-line rtcm.cpp into per-role TUs
- rtcm_internal.hpp: former anonymous-namespace bit/CRC/message helpers
  as inline functions in libgnss::io::rtcm_internal
- rtcm_encode.cpp: MSM observation/ephemeris encoding (~1k lines)
- rtcm_decode_observation.cpp: MSM observation decoding
- rtcm_decode_nav.cpp: ephemeris and SSR decoding
- rtcm_reader.cpp: RTCMReader transports/reconnect + rtcm_utils
- rtcm.cpp keeps RTCMProcessor state, dispatchers, setBits/parseHeader

Verified: ctest 71/71, byte-identical rtk_positioning output on PPC
tokyo/run1, MADOCALIB parity suite 17/17.
